Update module regex pattern, allow recusrive modules
Allowing recursive modules makes it a lot easier to combine exisinting
modules that can't be fully "partialed", which is what I needed to do
for the regex modules since they depend on a supplied pattern source.

This also introduces a concept of a module getter pattern to help
distinguis between modules that can just be used directly to build
schemas and those that have to be evaluated first by supplying an
argument. The type of a partialed function isn't very user friendly,
which is what the default would have been without this.