Introduce multi metric concept
This lets you generate multiple series from a single metric, which
offers potentially substantial perf advantages for certain use cases.
This doesn't really map to whylogs so there is an inefficient wrapper
that I threw in that can automatically adapt the metric still. This also
ensures that metric names are declared upfront in the metric definition
so we can find out about them before we evaluate the metric, and perform
certain validations/catch mistaktes.