Added all of the metrics to the library, new lazy utility
This adds all of the metrics to the library of metrics so they can be
discovered. It also introduces a DynamicLazyInit class that is capable
of doing lazy initialization using functions that take arguments, and
maintains a cache of argument->item rather than just a single item, like
the LazyInit utility. This is useful for doing initialization with
models that require some input before creation.