Update the automatic whylogs schema gen to fix warnings
There were a lot of warnings before because I put the NO_FI_RESOLVERS in
once for each metric, which is apparently frowned upon. Now it's only
used once in the entire conversion and the only metrics that have
something unique added are the frequent items ones, and they use metric
names instead of types so there is no duplication/conflict.