Renames, support for input ids, and callback library
This introduces a library concept for callbacks like we have for the
metrics, with a few reasonable default callbacks in there for sending
validation failures to slack or any arbitrary webhook with some common
auth options.

This also lets you tag along an `id` column with your input which will
be used in validators and returned in the final output. Before, we just
generated auto increment ids (through pandas index).