Address themes feedback, fix multi metric whylogs converstion
This addresses the feedback around the jailbreak/refusal metrics and
also deletes the lame converstion scheme for the whylogs schemas that
ran the langkit multi metrics n times and uses the whylogs method of
delivering multiple metric outputs instead. A no-op in terms of
interface and output on the langkit side but performs much better under
the hood, which is important now that we have a standard metric that
emits multiple output series (presidio pii).