Only download nltk data if it isn't already downloaded
The most recent version of nltk makes network requests even when data is
downloaded locally, perhaps to see if there is a newer version. This
makes it hard to assert that we have everything downloaded in CI though
because we expect no further network requests to happen. This adds an
explicit check for the download before using the more generic .download
method to avoid those network requests and output.