Add additional cache_asset functions, share single transformer cache logic
There were metrics that weren't using the utility cached transformer
method despite using the same transformer. This removes the duplicated
references to that transformer creation, as well as adds a cache_assets
hook for the themes module so it behaves when we use the
cache_assets=False flag through workflows.