Add token count metric, fix edge case in metric evaluator, update recommended
This updates the recommended set of metrics based on some performance
tests. It gets rid of a few things on the prompt that people probably
wouldn't care as much about (toxicity/sentiment) that were expensive,
and adds in token count to both the prompt/response, which is pretty
fast and generally the base for billing.