Add a context concept to the workflow
This adds a new concept of a "context" and accompanying types for
populating that context. Metrics declare which context dependencies they
need and the workflow ensures that they're run.

The driver for this now is the lack of visibility into perf. Before this
change, metrics that all needed the same embeddings to be computed would
avoid wasted work recomputing them indirectly through @cache on the
types that generate the embeddings. This meant that the first metric
that happened to need the metric ended up looking pretty bad in terms of
perf with the rest looking good.

The context formalizes this and avoid the caching by pulling common
dependencies out and computing them once for everyone. If multiple
metrics declare a dependency on something then those are deduped as
well.